+++ packages/workflow/test/utils.test.ts
@@ -93,7 +93,7 @@ describe('jsonParse', () => {
 		expect(jsonParse('{ "a": 1 }')).toEqual({ a: 1 });
 	});
 
-	it('optionally throws `errorMessage`', () => {
+	it('optionally throws `errorMessage', () => {
 		expect(() => {
 			jsonParse('', { errorMessage: 'Invalid JSON' });
 		}).toThrow('Invalid JSON');
@@ -101,10 +101,6 @@ describe('jsonParse', () => {
 
 	it('optionally returns a `fallbackValue`', () => {
 		expect(jsonParse('', { fallbackValue: { foo: 'bar' } })).toEqual({ foo: 'bar' });
-	});
-
-	it('optionally returns a `fallbackValue` from a function', () => {
-		expect(jsonParse('', { fallbackValue: () => ({ foo: 'bar' }) })).toEqual({ foo: 'bar' });
 	});
 });
